Minor Refactoring.
