+++ config/grid.yaml
@@ -18,6 +18,9 @@ services:
   Pimcore\Bundle\StudioBackendBundle\Grid\Service\ConfigurationServiceInterface:
     class: Pimcore\Bundle\StudioBackendBundle\Grid\Service\ConfigurationService
 
+  Pimcore\Bundle\StudioBackendBundle\Grid\Service\SaveConfigurationServiceInterface:
+    class: Pimcore\Bundle\StudioBackendBundle\Grid\Service\SaveConfigurationService
+
   Pimcore\Bundle\StudioBackendBundle\Grid\Service\SystemColumnServiceInterface:
     class: Pimcore\Bundle\StudioBackendBundle\Grid\Service\SystemColumnService
